Develop

+++ ClaimProcessing.Application/Packagings/Commands/DeletePackaging/DeletePackagingCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using ClaimProcessing.Application.Common.Interfaces;
+﻿using ClaimProcessing.Application.Common.Exceptions;
+using ClaimProcessing.Application.Common.Interfaces;
 using ClaimProcessing.Domain.ValueObjects;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
@@ -14,12 +15,16 @@ namespace ClaimProcessing.Application.Packagings.Commands.DeletePackaging
         }
         public async Task Handle(DeletePackagingCommand request, CancellationToken cancellationToken)
         {
-            var packaging = await _context.Packagings.Where(p => p.Id == request.PackagingId).FirstOrDefaultAsync(cancellationToken);
+            var packaging = await _context.Packagings.Where(p => p.StatusId != 0 && p.Id == request.PackagingId).FirstOrDefaultAsync(cancellationToken);
+
+            if (packaging == null)
+            {
+                throw new NullException(request.PackagingId);
+            }
+
             var dimensions = new Dimensions(packaging.Dimensions.Height, packaging.Dimensions.Width, packaging.Dimensions.Depth);
-            
             _context.Packagings.Remove(packaging);
             packaging.Dimensions = dimensions;
-
             await _context.SaveChangesAsync(cancellationToken);
         }
     }
